spring-boot集成redis,并使用redis缓存

+++ src/main/java/com/crw/service/SbpProductService.java
@@ -4,7 +4,7 @@ import com.crw.model.SbpProduct;
 
 public interface SbpProductService {
 
-    Integer insert(SbpProduct sbpProduct);
+    Long insert(SbpProduct sbpProduct);
 
     boolean update(SbpProduct sbpProduct);
